requirements.txt
asgiref==3.7.2
Django==4.2.7
sqlparse==0.4.4
tzdata==2023.3
django-ckeditor==6.7.0
